Revert "update: phaseUpdate.notification 내부 로직 함수화 및 분리"
This reverts commit 960085a341473b0bb9db78174d754c045aedc77e.
